api/requirements.txt
aioredis==2.0.0
anyio==3.3.4
asgiref==3.4.1
async-timeout==4.0.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==35.0.0
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.70.0
gunicorn==20.1.0
h11==0.12.0
httptools==0.2.0
idna==3.3
mergedeep==1.3.4
motor==2.5.1
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pymongo==3.12.1
python-dotenv==0.19.1
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.26.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.0
